Refactor ssnowflake profile mapping
Introduce the SnowflakeBaseProfileMapping class with behaviour that should be shared between multiple Snowflake profile mapping classes
